more warnings patches from Michael D. Adams
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1538 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/csstrings.h
@@ -36,7 +36,7 @@
 extern char temporary[256];
 #define ptemporary (*(Str255 *)temporary)
 
-extern short countstr(
+extern size_t countstr(
 	short resid);
 
 extern unsigned char *getpstr(
